+++ src/components/common/index.tsx
@@ -3,4 +3,4 @@ export { default as SpinnerButton } from './SpinnerButton';
 export { default as PrimaryModal } from './PrimaryModal';
 export { default as LoadingBar } from './LoadingBar';
 
-export { default as CancleButton } from './CancleButton';
+export { default as CancelButton } from './CancelButton';
